Move plugins to core/plugins
