+++ src/core/plugins/EnforceNonZeroErrorPlugin.d.ts
