+++ src/core/plugins/ImplicitTilingPlugin.js
