+++ src/core/plugins/SUBTREELoader.js
@@ -2,9 +2,9 @@
  * Structure almost identical to Cesium, also the comments and the names are kept
  * https://github.com/CesiumGS/cesium/blob/0a69f67b393ba194eefb7254600811c4b712ddc0/packages/engine/Source/Scene/Implicit3DTileContent.js
  */
-import { LoaderBase } from '../../core/renderer/loaders/LoaderBase.js';
-import { readMagicBytes } from '../../core/renderer/utilities/readMagicBytes.js';
-import { arrayToString } from '../../core/renderer/utilities/arrayToString.js';
+import { LoaderBase } from '../renderer/loaders/LoaderBase.js';
+import { readMagicBytes } from '../renderer/utilities/readMagicBytes.js';
+import { arrayToString } from '../renderer/utilities/arrayToString.js';
 
 function isOctreeSubdivision( tile ) {
