+++ src/core/plugins/auth/CesiumIonAuth.js
