+++ src/core/plugins/auth/GoogleCloudAuth.js
@@ -1,4 +1,4 @@
-import { traverseSet } from '../../../core/renderer/tiles/traverseFunctions.js';
+import { traverseSet } from '../../renderer/tiles/traverseFunctions.js';
 
 const TILES_MAP_URL = 'https://tile.googleapis.com/v1/createSession';
