+++ src/core/plugins/index.d.ts
@@ -1,0 +1,2 @@
+export * from './ImplicitTilingPlugin.js';
+export * from './EnforceNonZeroErrorPlugin.js';
