+++ src/core/plugins/loaders/QuantizedMeshLoaderBase.js
@@ -1,4 +1,4 @@
-import { LoaderBase } from '../../../core/renderer/loaders/LoaderBase.js';
+import { LoaderBase } from '../../renderer/loaders/LoaderBase.js';
 
 export function zigZagDecode( value ) {
